Update OSMIndoorLayer.java (#365)
The get level method does not parse levels with double type i.e. 0.5.
The app crashes at the time of loading. The solution to this issue, is
parsing double and then converting it into integer value.
e.g. level 0.5 to level 0

+++ vtm-jeo/src/org/oscim/layers/OSMIndoorLayer.java
@@ -172,7 +172,7 @@ public class OSMIndoorLayer extends JeoVectorLayer {
 
         o = f.get("level");
         if (o instanceof String) {
-            return Integer.parseInt((String) o);
+            return (int) Double.parseDouble((String) o);
         }
 
         return 0;
